Split unit tests from incremental.rs into incremental_test.rs
Move 12 test functions (~370 lines) from the inline test module in
incremental.rs to a dedicated incremental_test.rs file. This reduces
the source file size and improves rust-analyzer performance.

Snapshot files are moved to match the new module path.

Partial fix for #5116
